feat: refactor(table): enhance zeta-table styles and structure
- Updated styles for sortable table headers to use button elements for better accessibility and interaction.
- Consolidated type exports in table.ts for improved clarity and organization.
- Introduced a new table.types.ts file to define table-related types, enhancing type management.
- Refined utility functions for consistent handling of cell values and event details in tests.
- Improved test assertions for event detail consistency across table feature and interaction tests.
